inscripciones_soap/requirements.txt
spyne
mysql-connector-python
lxml
